added explicit fp32, ran ruff

+++ src/weathergen/model/model.py
@@ -561,7 +561,7 @@ class Model(torch.nn.Module):
 
             # embed token coords, concatenating along batch dimension (which is taking care of through
             # the varlen attention)
-            with torch.amp.autocast("cuda", enabled=False):
+            with torch.amp.autocast("cuda", dtype=torch.float32, enabled=False):
                 tc_tokens = torch.cat(
                     [
                         checkpoint(
